pyproject.toml
[project]
name = "xm-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.4.8",
    "aiosqlite>=0.21.0",
    "fastapi>=0.128.7",
    "loguru>=0.7.3",
    "openai>=2.20.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy==2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
